refactor: Split window tests.rs into focused test modules
Split the 839-line window tests.rs file into smaller, focused modules for better organization and maintainability:

- partitions.rs: Partition-related tests (2 tests)
- frames.rs: Frame calculation tests (5 tests)
- ranking.rs: Ranking function tests - ROW_NUMBER, RANK, DENSE_RANK, NTILE (15 tests)
- aggregates.rs: Aggregate window function tests - COUNT, SUM, AVG, MIN, MAX (16 tests)
- lag_lead.rs: LAG/LEAD value function tests (8 tests)

Benefits:
- Improved test organization and navigation
- Easier to locate specific test cases
- Better parallel test execution potential
- Clearer test categorization by functionality

All 46 tests pass successfully.

Closes #437
