+++ crates/executor/src/evaluator/window/tests/mod.rs
@@ -1,0 +1,8 @@
+use super::*;
+use crate::evaluator::window::utils::evaluate_expression;
+
+mod partitions;
+mod frames;
+mod ranking;
+mod aggregates;
+mod lag_lead;
